integration-tests/requirements.txt
behave==1.*
black
dictdiffer
draymed==2.*
environs==4.*
faker==4.*
isort
jsonpatch==1.*
kombu==4.*
mypy
neo4j-driver==1.*
pycryptodomex==3.*
python-jose==3.*
reportportal-behave-client==1.*
requests==2.*
she-data-generation>=1.0.5
she-logging
types-requests
